requirements.txt
coverage
python-dateutil
pytest
pytest-recording
requests
ruff